Change time_to_respawn from int to float

+++ src/state/include/actor/actor.h
@@ -81,7 +81,7 @@ protected:
 	/**
 	 * The respawn time remaining (0 if alive)
 	 */
-	int64_t time_to_respawn;
+	float time_to_respawn;
 	/**
 	 * For base poisoning calculation
 	 * If the actor spends too much time near the base, he starts to lose hp
@@ -145,7 +145,7 @@ public:
 		int64_t max_speed,
 		int64_t size,
 		int64_t total_respawn_time,
-		int64_t time_to_respawn,
+		float time_to_respawn,
 		int64_t time_spent_near_base,
 		physics::Vector2D position,
 		physics::Vector2D velocity,
